getTextMatches in search now finds prepositions
Because mongodb text search/indexes skip stop words such as 'and' and 'from', most of the prepositions cannot be found using search. Now, if a text search results in no matches, words with the exact english are searched for and returned.
An index has been added for 'english' in the WordSchema to speed up these searches